+++ cardano-cli/src/Cardano/CLI/EraBased/Governance/Option.hs
@@ -5,11 +5,10 @@ module Cardano.CLI.EraBased.Governance.Option
   ( GovernanceCmds (..)
   , renderGovernanceCmds
   , pGovernanceCmds
-  , pCreateMirCertificatesCmds
   )
 where
 
-import Cardano.Api (Convert (..), ShelleyBasedEra, ShelleyToBabbageEra, forShelleyBasedEraMaybeEon)
+import Cardano.Api (Convert (..))
 import Cardano.Api.Experimental qualified as Exp
 
 import Cardano.CLI.EraBased.Common.Option
@@ -18,9 +17,7 @@ import Cardano.CLI.EraBased.Governance.Command
 import Cardano.CLI.EraBased.Governance.Committee.Option
 import Cardano.CLI.EraBased.Governance.DRep.Option
 import Cardano.CLI.EraBased.Governance.Vote.Option
-import Cardano.CLI.Parser
 
-import Data.Foldable
 import Options.Applicative
 import Options.Applicative qualified as Opt
 
@@ -44,61 +41,3 @@ pGovernanceCmds =
     , fmap GovernanceDRepCmds <$> pGovernanceDRepCmds (convert Exp.useEra)
     , fmap GovernanceVoteCmds <$> pGovernanceVoteCmds (convert Exp.useEra)
     ]
-
-pCreateMirCertificatesCmds :: ShelleyBasedEra era -> Maybe (Parser (GovernanceCmds era))
-pCreateMirCertificatesCmds era = do
-  w <- forShelleyBasedEraMaybeEon era
-  pure $
-    Opt.hsubparser $
-      commandWithMetavar "create-mir-certificate" $
-        Opt.info (pMIRPayStakeAddresses w <|> mirCertParsers w) $
-          Opt.progDesc "Create an MIR (Move Instantaneous Rewards) certificate"
-
-mirCertParsers
-  :: ()
-  => ShelleyToBabbageEra era
-  -> Parser (GovernanceCmds era)
-mirCertParsers w =
-  asum
-    [ Opt.hsubparser $
-        commandWithMetavar "stake-addresses" $
-          Opt.info (pMIRPayStakeAddresses w) $
-            Opt.progDesc "Create an MIR certificate to pay stake addresses"
-    , Opt.hsubparser $
-        commandWithMetavar "transfer-to-treasury" $
-          Opt.info (pGovernanceCreateMirCertificateTransferToTreasuryCmd w) $
-            Opt.progDesc "Create an MIR certificate to transfer from the reserves pot to the treasury pot"
-    , Opt.hsubparser $
-        commandWithMetavar "transfer-to-rewards" $
-          Opt.info (pGovernanceCreateMirCertificateTransferToReservesCmd w) $
-            Opt.progDesc "Create an MIR certificate to transfer from the treasury pot to the reserves pot"
-    ]
-
-pMIRPayStakeAddresses
-  :: ()
-  => ShelleyToBabbageEra era
-  -> Parser (GovernanceCmds era)
-pMIRPayStakeAddresses w =
-  GovernanceCreateMirCertificateStakeAddressesCmd w
-    <$> pMIRPot
-    <*> some (pStakeAddress Nothing)
-    <*> some pRewardAmt
-    <*> pOutputFile
-
-pGovernanceCreateMirCertificateTransferToTreasuryCmd
-  :: ()
-  => ShelleyToBabbageEra era
-  -> Parser (GovernanceCmds era)
-pGovernanceCreateMirCertificateTransferToTreasuryCmd w =
-  GovernanceCreateMirCertificateTransferToTreasuryCmd w
-    <$> pTransferAmt
-    <*> pOutputFile
-
-pGovernanceCreateMirCertificateTransferToReservesCmd
-  :: ()
-  => ShelleyToBabbageEra era
-  -> Parser (GovernanceCmds era)
-pGovernanceCreateMirCertificateTransferToReservesCmd w =
-  GovernanceCreateMirCertificateTransferToReservesCmd w
-    <$> pTransferAmt
-    <*> pOutputFile
